Rename extensions/quantum_initializer/_initializer
