+++ qiskit/extensions/quantum_initializer/initializer.py
